DeezyMatch saves models for each epoch, i.e., checkpoints (unless `validation` option in the input file is not equal to 1)
When DeezyMatch finishes the last epoch, it will save the model with least validation loss as well. Morevoer, DeezyMatch has an `early stopping` functionality. This can be activated by setting the `early_stopping_patience` option in the input file. This option specifies the number of epochs with no improvement after which training will be stopped and the model with the least validation loss will be saved.